Fixed problem with HMR not working
