+++ .changeset/dull-kings-sparkle.md
@@ -2,6 +2,7 @@
 "@kuma-ui/system": minor
 "@kuma-ui/sheet": minor
 "@kuma-ui/vite": minor
+"@kuma-ui/compiler": minor
 ---
 
 fix/kuma.config can hmr
