+++ packages/compiler/src/extractor/extract.ts
@@ -1,25 +1,12 @@
-import {
-  Project,
-  Node,
-  SyntaxKind,
-  JsxOpeningElement,
-  JsxSelfClosingElement,
-} from "ts-morph";
-import {
-  isStyledProp,
-  isPseudoProps,
-  normalizePseudo,
-  all,
-  SystemStyle,
-  StyleGenerator,
-} from "@kuma-ui/system";
+import { Node, JsxOpeningElement, JsxSelfClosingElement } from "ts-morph";
+import { isStyledProp, isPseudoProps, StyleGenerator } from "@kuma-ui/system";
 import {
   componentDefaultProps,
   componentList,
   isComponentProps,
   componentHandler,
 } from "@kuma-ui/core/components/componentList";
-import { theme } from "@kuma-ui/sheet";
+import { Tokens, theme } from "@kuma-ui/sheet";
 import { stableStringify } from "../utils/stableStringify";
 
 export const extractProps = (
@@ -49,7 +36,21 @@ export const extractProps = (
     ...defaultProps,
     ...propsMap,
   })) {
-    if (isStyledProp(propName.trim())) {
+    if (
+      typeof propValue === "string" &&
+      /[a-zA-Z]+\.[a-zA-Z0-9]+/.test(propValue)
+    ) {
+      const userTheme = theme.getUserTheme();
+      const propKey = propValue.split(".")[0] as Tokens;
+      if (userTheme[propKey] !== undefined) {
+        for (const key in userTheme[propKey]) {
+          if (propValue.trim() === key) {
+            styledProps[propName.trim()] = userTheme[propKey]![key];
+            break;
+          }
+        }
+      }
+    } else if (isStyledProp(propName.trim())) {
       styledProps[propName.trim()] = propValue;
     } else if (isPseudoProps(propName.trim())) {
       pseudoProps[propName.trim()] = propValue;
